cleanup

+++ java/socket/src/msgs/GenericDeserializer.java
@@ -9,7 +9,7 @@ public class GenericDeserializer {
 		Hash hash = new Hash(256); 
 		hash.deserialize(_in);
 		int  type = _in.readInt();
-		int  size = (int)_in.readLong();
+		int  size = _in.readInt();
 		
 		SocketMsg msg = null;
 		if(DataType.isSequence(type, size)) {
